+++ src/main/java/pl/vertty/arivi/guilds/task/RegenerationTask.java
@@ -22,7 +22,7 @@ public class RegenerationTask extends NukkitRunnable
             final BlockRegeneration blockRegeneration = this.blocks.get(0);
             if (blockRegeneration != null) {
                 final Location location = blockRegeneration.getLocation();
-                location.getLevel().setBlock(new Vector3((double)location.getFloorX(), (double)location.getFloorY(), (double)location.getFloorZ()), Block.get(blockRegeneration.getIdBlock()));
+                location.getLevel().setBlock(new Vector3(location.getFloorX(), location.getFloorY(), location.getFloorZ()), Block.get(blockRegeneration.getIdBlock()));
             }
             this.blocks.remove(blockRegeneration);
             if (this.blocks.isEmpty()) {
